maintain: move user api tests to their own file
